Support download cache when building >= 1.5
Instead of downloading 1.4 behind bitbake's back, list 1.4 as part of >=
1.5's sources, and go from there.

I ran into a weird situation with 1.5 where it would not compile the
compiler with CGO_ENABLED=1 unless CC_FOR_TARGET is a single word (i.e.
it cannot be "gcc --sysroot=..." or some such)

1.4 is likely broken.

Needs to provide a "go" class to set GOARCH and friends correctly.