Install Go compiler for go-native
This is a workaround: when building go-cross, two binaries are
installed, one for the host and one for the target. When building
go-native, only the binary for the host is installed. The patch is meant
to remove the target binary, but it's causing the host binary to be
missing in go-native.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@gmail.com>