Reorganize go recipes
Instead of having one bb file for cross and native packages, have a
single go bb file that uses BBCLASSEXTEND to add cross and native
packages.

To make this work, rename go-bootstrap to go-bootstrap-native.

Put all the patches in a single go directory.

Remove one of the intermediate .inc files.

Move the license checksum to the version-specific file since Go 1.7
chnages the text of the license.

As a side-effect, support a go package on the target, in case you need
to Go compiler in the target system.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@gmail.com>