Add GO_INSTALL variable to go.bbclass
If you provide a value for GO_INSTALL, instead of trying to install
${GO_IMPORT}/... the go_compile and go_install tasks will limit
themselves to working with the specified packages and their
dependencies.

This provides supports for repositories where there's more code that
what you normally want to build (examples) or repositories that contain
their main functionality in a subdirectory (e.g. a "cmd/foo" directory
with the actual program that you want).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@gmail.com>